requirements.txt
Django==1.10.3
psycopg2==2.6.2
Pillow==3.4.2
djangorestframework==3.5.3
django-autofixture==0.12.1
django-oauth-toolkit==0.10.0
django-filter==1.0.0
djangorestframework-gis==0.11
django-push-notifications==1.4.1
redis==2.10.5
celery==3.1.24
drfdocs==0.0.11
django-imagekit==3.3
